refactor(workflow): split 1204-line module into 8 cohesive submodules
oh-my-symphony 0.6.4's workflow.py had grown past the 800-line cap and
mixed five different concerns: front-matter parsing, value coercion,
frozen typed config dataclasses, the long YAML->ServiceConfig builder,
dispatch preflight validation, and the hot-reload holder. Splitting
along those existing seams without changing behavior:

  workflow/
    constants.py    65 — defaults, env keys, _VAR_PATTERN
    parser.py       92 — WorkflowDefinition + parse/load/resolve
    coercion.py    125 — $VAR / ~ expand + _as_*/normalize/path helpers
    config.py      359 — frozen dataclass config types
    builder.py     599 — build_service_config + strict validators
    preflight.py    75 — validate_for_dispatch (dispatch-time gate)
    state.py        47 — WorkflowState (§6.2 hot reload)
    __init__.py    141 — re-exports preserving every public + private
                         dotted name tests rely on

Public API and dotted-path stub targets unchanged — every symbol that
used to be importable from `symphony.workflow` still is. Full pytest
suite green: 565 passed, 6 skipped (identical to baseline).
